move struct_writers to code_writers directory
