+++ swiftwinrt/CMakeLists.txt
@@ -52,7 +52,9 @@ target_sources(swiftwinrt PUBLIC
     types.cpp
     metadata_filter.cpp
     code_writers/type_writers.cpp
-    resources.rc)
+    code_writers/struct_writers.cpp
+    resources.rc
+)
 
 # Make resources.rc depend on the files it embeds
 file(GLOB_RECURSE SUPPORT_FILES CONFIGURE_DEPENDS Resources/*)
