+++ swiftwinrt/code_writers.h
@@ -1,190 +1,10 @@
 #pragma once
+#include "code_writers/common_writers.h"
 #include "code_writers/type_writers.h"
+#include "code_writers/can_write.h"
 #include "metadata_cache.h"
 namespace swiftwinrt
 {
-    struct finish_with
-    {
-        writer& w;
-        void (*finisher)(writer&);
-
-        finish_with(writer& w, void (*finisher)(writer&)) : w(w), finisher(finisher) {}
-        finish_with(finish_with const&) = delete;
-        void operator=(finish_with const&) = delete;
-
-        ~finish_with() { finisher(w); }
-    };
-
-    static void write_nothing(writer&)
-    {
-    }
-
-    template <typename T>
-    inline void write_type_mangled(writer& w, T const& type)
-    {
-        auto push_mangled = w.push_mangled_names(true);
-        w.write(type);
-    }
-
-    template <typename T>
-    auto bind_type_mangled(T const& type)
-    {
-        return [&](writer& w)
-        {
-            write_type_mangled(w, type);
-        };
-    }
-
-    template <typename T>
-    inline void write_type_abi(writer& w, T const& type)
-    {
-        auto push_abi = w.push_abi_types(true);
-        w.write(type);
-    }
-
-    template <typename T>
-    auto bind_type_abi(T const& type)
-    {
-        return [&](writer& w)
-        {
-            write_type_abi(w, type);
-        };
-    }
-
-    template<typename T>
-    inline void write_generic_impl_name(writer& w, T const& type)
-    {
-        // for IReference<> types we use the same IPropertyValueImpl class that is
-        // specially generated. this type can hold any value type and implements
-        // the appropriate interface
-        if (is_winrt_ireference(type))
-        {
-            w.write("%.%", impl_namespace("Windows.Foundation"), "IPropertyValueImpl");
-        }
-        else
-        {
-            std::string implName;
-            {
-                auto use_mangled = w.push_mangled_names(true);
-                implName = w.write_temp("%Impl", type);
-            }
-
-            if (w.impl_names)
-            {
-                w.write(implName);
-            }
-            else 
-            {
-                // generics aren't public and so we use the namespace of the writer
-                w.write("%.%", impl_namespace(w.type_namespace), implName);
-            }
-        }
-    }
-
-    template<typename T>
-    inline void write_impl_name(writer& w, T const& type)
-    {
-        if (is_generic_inst(type))
-        {
-            write_generic_impl_name(w, type);
-        }
-        else
-        {
-            assert(!is_generic_def(type));
-
-            type_name type_name{ type };
-            std::string implName = w.write_temp("%Impl", type_name.name);
-
-            if (w.type_namespace != type_name.name_space || w.mangled_names || w.full_type_names)
-            {
-                w.write("%.%", impl_namespace(type_name.name_space), implName);
-            }
-            else if (w.impl_names)
-            {
-                w.write(implName);
-            }
-            else
-            {
-                w.write("%.%", impl_namespace(type_name.name_space), implName);
-            }
-        }
-    }
-
-    template <typename T>
-    auto bind_impl_name(T const& type)
-    {
-        return [&](writer& w)
-        {
-            write_impl_name(w, type);
-        };
-    }
-
-    template <typename T>
-    auto bind_impl_fullname(T const& type)
-    {
-        return [&](writer& w)
-        {
-            auto full_name = w.push_full_type_names(true);
-            write_impl_name(w, type);
-        };
-    }
-
-    template<typename T>
-    inline void write_wrapper_name(writer& w, T const& type)
-    {
-        type_name type_name(type);
-
-        if (is_generic_inst(type))
-        {
-            auto mangled_name = w.push_mangled_names(true);
-            auto handlerWrapperTypeName = w.write_temp("%Wrapper", type);
-            if (w.full_type_names)
-            {
-                w.write("%.%", abi_namespace(w.type_namespace), handlerWrapperTypeName);
-            }
-            else
-            {
-                w.write(handlerWrapperTypeName);
-            }
-        }
-        else
-        {
-            assert(!is_generic_def(type));
-            
-            auto handlerWrapperTypeName = w.write_temp("%Wrapper", type_name.name);
-            if (w.full_type_names)
-            {
-                w.write("%.%", abi_namespace(type_name.name_space), handlerWrapperTypeName);
-            }
-            else
-            {
-                w.write(handlerWrapperTypeName);
-            }
-        }
-
-    }
-
-    template <typename T>
-    auto bind_wrapper_name(T const& type)
-    {
-        return [&](writer& w)
-        {
-            write_wrapper_name(w, type);
-        };
-    }
-
-    template <typename T>
-    auto bind_wrapper_fullname(T const& type)
-    {
-        return [&](writer& w)
-        {
-            auto full_type_names = w.push_full_type_names(true);
-            write_wrapper_name(w, type);
-        };
-    }
-
-    static bool can_write(writer& w, typedef_base const& type);
-
     static void write_enum_def(writer& w, enum_type const& type)
     {
         // Async status is defined in it's own header with a nice name of AsyncStatus.
@@ -228,178 +48,6 @@ namespace swiftwinrt
             }
         }
         w.write("}\n\n");
-    }
-
-    static bool can_write(generic_inst const& type);
-    static bool can_write(writer& w, const metadata_type* type)
-    {
-        if (auto typed = dynamic_cast<const typedef_base*>(type))
-        {
-            return can_write(w, *typed);
-        }
-        if (auto generics = dynamic_cast<const generic_inst*>(type))
-        {
-            return can_write(*generics);
-        }
-        if (auto mapped = dynamic_cast<const mapped_type*>(type))
-        {
-            // TODO: https://linear.app/the-browser-company/issue/WIN-103/swiftwinrt-write-iasyncinfo
-            if (mapped->type().TypeName() == "IAsyncInfo") return false;
-        }
-        return true;
-    }
-
-    static bool can_write(writer& w, TypeDef const& type)
-    {
-        return can_write(w, &w.cache->find(type.TypeNamespace(), type.TypeName()));
-    }
-
-    static bool can_write(generic_inst const& type)
-    {
-        // TODO: WIN-275: Code generation for nested generics
-        for (auto genarg : type.generic_params())
-        {
-            if (is_generic_inst(genarg))
-            {
-                return false;
-            }
-        }
-
-        auto name = type.generic_type_abi_name();
-        return name == "IReference" || name == "IEventHandler" || name == "ITypedEventHandler"
-            || name == "IVector" || name == "IVectorView" || name == "IMap" || name == "IMapView";
-    }
-
-    static bool can_write(writer& w, function_def const& function, bool allow_special = false)
-    {
-        auto method = function.def;
-
-        // Don't support writing specials (events/properties) unless told to do so (i.e. for vtable)
-        if (method.SpecialName() && !allow_special) return false;
-
-        auto method_name = get_swift_name(method);
-
-        // TODO: WIN-30 swiftwinrt: support async/await
-        if (function.is_async()) return false;
-        for (auto& param: function.params)
-        {
-            auto param_name = get_swift_name(param);
-
-            // TODO: WIN-32 swiftwinrt: support array types
-            if (param.signature.Type().is_szarray() ||
-                param.signature.Type().is_array())
-            {
-                return false;
-            }
-
-            // TODO: support reference parameters
-            if (param.signature.ByRef() && is_guid(get_category(param.type)))
-            {
-                return false;
-            }
-
-            if (!can_write(w, param.type))
-            {
-                return false;
-            }
-        }
-        
-        if (function.return_type)
-        {
-            auto returnType = function.return_type.value();
-            if (returnType.signature.Type().is_array() ||
-                returnType.signature.Type().is_szarray())
-            {
-                return false;
-            }
-            if (!can_write(w, function.return_type.value().type))
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    static bool can_write(writer& w, property_def const& prop)
-    {
-        if (prop.getter)
-        {
-            return can_write(w, prop.getter.value(), true);
-        }
-        if (prop.setter)
-        {
-            return can_write(w, prop.setter.value(), true);
-        }
-        assert(false); // property should have at least one
-        return true;
-    }
-
-    static bool can_write_default(const std::vector<named_interface_info>& required_interfaces)
-    {
-        for (auto&& iface : required_interfaces)
-        {
-            // when getting the interfaces we populate them with the type name
-            // if we can't write the type name then we can't write the type and
-            // so we'll return false
-            if (iface.second.is_default)
-            {
-                // TODO: WIN-274 Code generation for IIterable/IIterator
-                // TODO: WIN-124 Code generation for IObservableVector and IObservableMap
-                auto name = iface.second.type->swift_full_name();
-                if (name.starts_with("Windows.Foundation.Collections.IIterable")
-                    || name.starts_with("Windows.Foundation.Collections.IIterator")
-                    || name.starts_with("Windows.Foundation.Collections.IObservableVector")
-                    || name.starts_with("Windows.Foundation.Collections.IObservableMap"))
-                {
-                    return false;
-                }
-
-                if (iface.first.empty()) return false;
-            }
-        }
-
-        return true;
-    }
-
-    static bool can_write(writer& w, typedef_base const& type)
-    {
-        auto typeName = get_full_type_name(type);
-        if (!w.filter.includes(type.type())) return false;
-
-        auto category = get_category(type.type());
-        if (category == category::enum_type) return true;
-
-        // TODO: WIN-65 swiftwinrt: support generic types
-        auto generics = type.type().GenericParam();
-        if (!empty(generics))
-        {
-            // don't write abi types bc that will put it in a UnsafeMutablePointer<>
-            auto non_mangled = w.push_mangled_names(false);
-            auto written = w.write_temp("%", type);
-            auto can_write = !written.empty();
-            if (!can_write)
-            {
-                return false;
-            }
-        }
-
-        if (auto iface = dynamic_cast<const interface_type*>(&type))
-        {
-            if (!can_write_default(iface->required_interfaces))
-            {
-                return false;
-            }
-        }
-        else if (auto classType = dynamic_cast<const class_type*>(&type))
-        {
-            if (!can_write_default(classType->required_interfaces))
-            {
-                return false;
-            }
-        }
-
-        return true;
     }
 
     static void write_guid_value(writer& w, std::vector<FixedArgSig> const& args)
@@ -656,87 +304,6 @@ namespace swiftwinrt
             s();
             auto param_name = function.return_type.value().name;
             w.write("&%", param_name);
-        }
-    }
-
-    static void write_consume_type(writer& w, metadata_type const* type, std::string_view const& name)
-    {
-        TypeDef signature_type{};
-        auto category = get_category(type, &signature_type);
-       
-        if (category == param_category::array_type)
-        {
-            // TODO: WIN-32 swiftwinrt: add support for array types
-            XLANG_ASSERT("**TODO: implement array type in write_consume_return_type");
-        }
-        else if (category == param_category::object_type)
-        {
-            if (is_interface(type))
-            {
-                w.write("%.unwrapFrom(abi: %)",
-                    bind_wrapper_fullname(type),
-                    name);
-            }
-            else if (is_delegate(type))
-            {
-                if (w.consume_types)
-                {
-                    w.write("_%", name);
-                }
-                else
-                {
-                    w.write("%.tryUnwrapFrom(abi: %)", bind_wrapper_name(type), name);
-                }
-            }
-            else if (is_class(type))
-            {
-                w.write(".from(abi: %)", name);
-            }
-            else
-            {
-                //TODO: implement generic object type
-                w.write(".from(%)", name);
-            }
-
-        }
-        else if (category == param_category::struct_type)
-        {
-            if (w.abi_types)
-            {
-                w.write(".from(swift: %)", name);
-            }
-            else
-            {
-                w.write(".from(abi: %)", name);
-            }
-        }
-        else if (is_type_blittable(category))
-        {
-            // fundamental types can just be simply copied to since the types match
-            w.write(name);
-        }
-        else if (w.abi_types && category == param_category::string_type)
-        {
-            auto format = "try! HString(%).detach()";
-            w.write(format, name);
-        }
-        else if (category == param_category::generic_type)
-        {
-            if (is_winrt_ireference(type))
-            {
-                w.write(".init(ref: %)", name);
-            }
-            else
-            {
-                w.write("%.unwrapFrom(abi: %)",
-                    bind_wrapper_fullname(type),
-                    name);
-            }
-        }
-        else
-        {
-            auto format = ".init(from: %)";
-            w.write(format, name);
         }
     }
 
@@ -2539,119 +2106,7 @@ private var _default: SwiftABI!
         write_class_impl(w, type);
     }
 
-    static void write_struct_initializer_params(writer& w, struct_type const& type)
-    {
-        separator s{ w };
-
-        for (auto&& field : type.members)
-        {
-            // WIN-64 - swiftwinrt: support boxing/unboxing
-            // WIN-65 - swiftwinrt: support generic types
-            if (can_write(w, field.type))
-            {
-                s();
-
-                w.write("%: %", get_swift_name(field), field.type);
-            }
-           
-        }
-    }
-
-    static void write_struct_init_extension(writer& w, struct_type const& type)
-    {
-        bool is_blittable = is_struct_blittable(type);
-        if (is_blittable)
-        {
-            w.write(R"(extension % {
-    public static func from(swift: %) -> % {
-        .init(%)
-    }
-}
-)", bind_type_mangled(type), get_full_swift_type_name(w,type), bind_type_mangled(type), bind([&](writer& w) {
-                    separator s{ w };
-                    for (auto&& field : type.members)
-                    {
-                        s();
-
-                        if (dynamic_cast<const struct_type*>(field.type))
-                        {
-                            w.write("%: .from(swift: swift.%)",
-                                get_abi_name(field),
-                                get_swift_name(field)
-                            );
-                        }
-                        else
-                        {
-                            w.write("%: swift.%",
-                                get_abi_name(field),
-                                get_swift_name(field)
-                            );
-                        }
-  
-                    }
-    }));
-        }
-    }
-
-    static void write_struct(writer& w, struct_type const& type)
-    {
-        w.write("public struct % {\n", type);
-        {
-            auto indent_guard1 = w.push_indent();
-            for (auto&& field : type.members)
-            {
-                auto field_type = field.type;
-                // WIN-64 - swiftwinrt: support boxing/unboxing
-                // WIN-65 - swiftwinrt: support generic types
-                if (!can_write(w, field_type)) continue;
-
-                w.write("public var %: %%\n", 
-                    get_swift_name(field),
-                    bind<write_type>(*field_type, write_type_params::swift_allow_implicit_unwrap),
-                    bind<write_default_init_assignment>(*field_type, projection_layer::swift));
-            }
-
-            w.write("public init() {}\n");
-            w.write("public init(%) {\n", bind<write_struct_initializer_params>(type));
-            {
-                auto indent_guard2 = w.push_indent();
-                for (auto&& field : type.members)
-                {
-                    // WIN-64 - swiftwinrt: support boxing/unboxing
-                    // WIN-65 - swiftwinrt: support generic types
-                    if (can_write(w, field.type))
-                    {
-                        auto field_name = get_swift_name(field);
-                        w.write("self.% = %\n", field_name, field_name);
-                    }
-                }
-            }
-            w.write("}\n");
-
-            w.write(R"(public static func from(abi: %) -> % {
-    .init(%)
-}
-)", bind_type_mangled(type),
-    type,
-    bind([&](writer& w) {
-                    separator s{ w };
-                    for (auto&& field : type.members)
-                    {
-                        if (can_write(w, field.type))
-                        {
-                            s();
-                            std::string from = std::string("abi.").append(get_abi_name(field));
-                            w.write("%: %",
-                                get_swift_name(field),
-                                bind<write_consume_type>(field.type, from)
-                            );
-                        }
-                    }
-    }));
-         
-        }
-        w.write("}\n\n");
-    }
+    
 
     template <typename T>
     static void write_query_interface_case(writer& w, T const& type, metadata_type const& iface, bool cast)
